Started to extract some steps
Tealised that the whole point is for them to be idempotent to be useful
